% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freegeomsf-package.R
\docType{package}
\name{freegeomsf-package}
\alias{freegeomsf-package}
\alias{_PACKAGE}
\title{freegeomsf: Free 'geom_sf' from its Aspect Ratio Shackles}
\description{
Scales sf geometry to simulate supporting free scales with ggplot2 plot facets.
}
\author{
\strong{Maintainer}: Michael Sumer \email{mdsumer@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
