% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freegeomsf.R
\name{freegeomsf}
\alias{freegeomsf}
\title{Rescale sf geometry for ggplot2 "free scales"}
\usage{
freegeomsf(x, ...)
}
\arguments{
\item{x}{sf object}

\item{...}{ignored for now}
}
\value{
sf object, with each element scale to unit square
}
\description{
Rescales sf geometries locally, to get around the preclusion of free scales
in ggplot2 faceting.
}
\details{
Because we can only simulate free scales, we can't carry the real world axes
to the plot (well it might be possible ...).

In future we might separate x and y, currently we rescale each geometry to the
unit square. We can't group the scaling atm, each one is done individually to its
own local range.
}
\examples{
data(sfx)
freegeomsf(sfx[1:15, ])
}
