% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freegeomsf-package.R
\docType{data}
\name{sfx}
\alias{sfx}
\title{A sf  object}
\description{
Obtained from 'CGAZ' a multipolygon dataset of New Zealand convert to single
polygons, with a field 'a' on them which is just row number. The data frame
is sorted by descending "area" (i.e. the size of the polygon, not the actual
geographic area, but I expect the order would be the same in this case, just
not generally).
}
\details{
See data-raw/ for the script to generate the sfx object.
}
